requirements.txt
# ============================================
# WOZ Backend - Python Dependencies
# ============================================

# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlmodel>=0.0.14
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
aiosqlite>=0.19.0

# Message Queue
aio-pika>=9.3.0

# Authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0

# PDF Generation
reportlab>=4.0.0

# Utilities
python-dotenv>=1.0.0
